move all kinds of functions from AngularKet to AngularKetBase
